# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SchoolCore - Multi-tenant school management platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /schools/{schoolId}/results:
    get:
      operationId: listResults
      tags: [results]
      summary: List academic results
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
        - name: classId
          in: query
          schema:
            type: integer
        - name: subjectId
          in: query
          schema:
            type: integer
        - name: studentId
          in: query
          schema:
            type: integer
        - name: term
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Results list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Result"
    post:
      operationId: createResult
      tags: [results]
      summary: Enter a result
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResultInput"
      responses:
        "201":
          description: Result created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
  /schools/{schoolId}/results/{resultId}:
    patch:
      operationId: updateResult
      tags: [results]
      summary: Update a result
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
        - name: resultId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResultUpdate"
      responses:
        "200":
          description: Updated result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
  /schools/{schoolId}/results/report:
    get:
      operationId: getResultsReport
      tags: [results]
      summary: Generate results report for a class/term
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
        - name: classId
          in: query
          schema:
            type: integer
        - name: term
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Results report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentResultReport"

  # ─── FEES ────────────────────────────────────────────────────────────────────
  /schools/{schoolId}/fees:
    get:
      operationId: listFees
      tags: [fees]
      summary: List fee records
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
        - name: studentId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [paid, pending, partial]
      responses:
        "200":
          description: Fee records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeeRecord"
    post:
      operationId: createFeeRecord
      tags: [fees]
      summary: Record a fee payment
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeeInput"
      responses:
        "201":
          description: Fee record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeeRecord"
  /schools/{schoolId}/fees/summary:
    get:
      operationId: getFeesSummary
      tags: [fees]
      summary: Fee collection summary
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Fees summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeesSummary"

  # ─── ANNOUNCEMENTS ───────────────────────────────────────────────────────────
...
  /parent/children/{studentId}/results:
    get:
      operationId: getChildResults
      tags: [parent]
      summary: Get child's academic results
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
        - name: term
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Child results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Result"
  /parent/children/{studentId}/attendance:
    get:
      operationId: getChildAttendance
      tags: [parent]
      summary: Get child's attendance
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Child attendance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChildAttendanceSummary"
  /parent/children/{studentId}/fees:
    get:
      operationId: getChildFees
      tags: [parent]
      summary: Get child's fee records
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Child fees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeeRecord"
...
components:
  schemas:
...
    AttendanceRecord:
      type: object
      required: [id, studentId, classId, date, status]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        classId:
          type: integer
        className:
          type: string
        date:
          type: string
          format: date
        status:
          type: string
          enum: [present, absent, late]
        schoolId:
          type: integer
...
    ChildAttendanceSummary:
      type: object
      required: [percentage, records]
      properties:
        percentage:
          type: number
        presentCount:
          type: integer
        absentCount:
          type: integer
        records:
          type: array
          items:
            $ref: "#/components/schemas/AttendanceRecord"

    # ─── RESULTS ───────────────────────────────────────────────────────────────
    Result:
      type: object
      required: [id, studentId, subjectId, score, grade, term]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        subjectId:
          type: integer
        subjectName:
          type: string
        classId:
          type: integer
        score:
          type: number
        grade:
          type: string
        comment:
          type: ["string", "null"]
        term:
          type: string
        academicYear:
          type: string
        schoolId:
          type: integer
    ResultInput:
      type: object
      required: [studentId, subjectId, classId, score, term]
      properties:
        studentId:
          type: integer
        subjectId:
          type: integer
        classId:
          type: integer
        score:
          type: number
        comment:
          type: ["string", "null"]
        term:
          type: string
        academicYear:
          type: string
    ResultUpdate:
      type: object
      properties:
        score:
          type: number
        comment:
          type: ["string", "null"]
        term:
          type: string
    StudentResultReport:
      type: object
      required: [studentId, studentName, results, average]
      properties:
        studentId:
          type: integer
        studentName:
          type: string
        results:
          type: array
          items:
            $ref: "#/components/schemas/Result"
        average:
          type: number
        rank:
          type: ["integer", "null"]

    # ─── FEES ──────────────────────────────────────────────────────────────────
    FeeRecord:
      type: object
      required: [id, studentId, amount, status, term, createdAt]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        amount:
          type: number
        amountPaid:
          type: number
        balance:
          type: number
        status:
          type: string
          enum: [paid, pending, partial]
        term:
          type: string
        description:
          type: ["string", "null"]
        schoolId:
          type: integer
        createdAt:
          type: string
          format: date-time
    FeeInput:
      type: object
      required: [studentId, amount, term]
      properties:
        studentId:
          type: integer
        amount:
          type: number
        amountPaid:
          type: number
        term:
          type: string
        description:
          type: ["string", "null"]
    FeesSummary:
      type: object
      required: [totalBilled, totalCollected, outstanding, paidCount, pendingCount]
      properties:
        totalBilled:
          type: number
        totalCollected:
          type: number
        outstanding:
          type: number
        paidCount:
          type: integer
        pendingCount:
          type: integer
        partialCount:
          type: integer

    # ─── ANNOUNCEMENTS ─────────────────────────────────────────────────────────
...